backend/requirements.txt
Django==3.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo[srv]==3.12.3
dnspython==2.4.2
Pillow==10.2.0
python-dotenv==1.0.1
drf-nested-routers==0.93.5
django-filter==23.5
python-dateutil==2.8.2
six==1.16.0
django-location-field==2.7.2
django-money==3.4.1
mongoengine==0.24.2 